requirements.txt
# requirements.txt

# =========================================================
# AIRFLOW CORE & PROVIDERS
# =========================================================
apache-airflow==2.9.0
apache-airflow-providers-postgres
apache-airflow-providers-http

# =========================================================
# DATABASE & DATA ENGINEERING
# =========================================================
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.52

# =========================================================
# AI / RAG / VECTOR SEARCH
# =========================================================
langchain==0.1.16
langchain-community==0.0.33
sentence-transformers==2.7.0
qdrant-client==1.9.0
openai==1.23.2

# =========================================================
# VALIDATION & CONFIG
# =========================================================
pydantic==1.10.14
python-dotenv==1.0.1

# =========================================================
# API & FRONTEND
# =========================================================
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
streamlit==1.32.2

# =========================================================
# BIG DATA 
# =========================================================
pyspark==3.5.1

# =========================================================
# CLOUD PROVIDERS
# =========================================================
azure-storage-blob==12.19.0